==> 02 Scripts/01. Orders Descriptive Analysis.ipynb <==
## Importing libraries

#Import libraries
import pandas as pd
import numpy as np
import os

## Importing data

#path to main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'
# import orders file
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

## Analysis

df.head()

df.tail()

df.shape

df.info()

df.columns

df.describe()

# import specific only specific columns and number of rows
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), nrows=10)
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)
df.head()
