==> 02 Scripts/02. Product Descriptive Analysis.ipynb <==
## Importing Libraries

#Importing libraries
import pandas as pd
import numpy as np
import os

## Importing Data

#path to main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'
#variable storing columns to keep from df
cols_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']
#import df
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = cols_list, index_col = False)
#import products file
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

## Analysis

#print first 5 rows of data from df
df.head()

#print prods first 5 rows 
df_prods.head()

#print prods last 5 rows
df_prods.tail(5)

#print prods names of cols
df_prods.columns

#print prods number of rows and columns
df_prods.shape

#prods stats
df_prods.describe()

#data type of department_id
df_prods.dtypes

01. What's the maximum value of the "aisle_id" column?
   134
02. What's the data type of the "department_id" column?
   integar (up to 64 characters)
