==> 02 Scripts/03. Data Wrangling and Subsetting.ipynb <==
## Importing Libraries

import pandas as pd
import numpy as nd
import os

## Importing data

#path to main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'
#variable storing columns to keep from df
cols_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

#import orders file
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = cols_list, index_col = False)

#import department file
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

#import products file
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

## Analysis

#count values for days since col
df_ords['days_since_prior_order'].value_counts(dropna = False)

#rename order_dow column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)
df_ords.head()

#change order_id to data type: string
df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['order_id'].dtype

#df_dep
df_dep.head()

#tranpose dep data
df_dep.T

df_dep_t = df_dep.T
df_dep_t

df_dep_t.reset_index()

#create new header
new_header = df_dep_t.iloc[0]
new_header

#create new df with only rows 1+ i.e. "delete" 1st row
df_dep_t_new = df_dep_t[1:]
df_dep_t_new

#set new header using new_header var
df_dep_t_new.columns = new_header
df_dep_t_new

#turn df_dept_t_new into a data dictionary
data_dict = df_dep_t_new.to_dict('index')
data_dict

df_prods.head()

print(data_dict.get('19'))

#create subset of snacks column
df_snacks = df_prods.loc[df_prods['department_id'].isin([19])]
df_snacks.head()

## Export Data

#export df
df_ords.to_csv(os.path.join(path,'02 Data', 'Prepared Data', 'orders_wrangled.csv'))

# 4.4 Tasks Steps  --------------------------------------

## df_ords

df_ords.head()

02. Find another identifier variable in the df_ords dataframe that doesn't need to be included in your analysis as a numeric variable and change it to a suitable format.

# change user_id to data type: string
df_ords['user_id'] = df_ords['user_id'].astype('int')
df_ords['user_id'].dtype

03. Look for a varable in your df_ords dataframe with an unituitive name and change its name without overwriting the dataframe.

# rename user_id to customer_id
# inplace = False so as to not overwrite the dataframe
df_ords.rename(columns = {'user_id' : 'customer_id'}, inplace = False)
df_ords.head()

04. Your client wants to know what the busiet hour is for placing orders. Find the frequency of the corresponding variable.

# show counts for each hour
df_ords['order_hour_of_day'].value_counts().sort_values()

# acording to the output below, hour 10 is the busiest hour of day

05. Determine the meaning behind a value of 4 in the 'department_id' column within the df_prods dataframe using a data dictionary.

# get the value of 4 from the data dictionary
print(data_dict.get('4'))

# it is the product department

06. The sales team in your client's organization wants to know more about breakfast item sales. Create a subset containing only the required information.

data_dict

# according to the dictionary printed 14 is the breakfast items
# creates subset for breakfood
df_breakfast = df_prods.loc[df_prods['department_id'].isin([14])]
df_breakfast.head()

07. They'd also like to see details about products that customers might use to throw dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood.

# 5: alcohol, 20:deli, 7: beverages, 12: meat/seafood
# creates subset for dinner party items lists in above comment
df_dinnerparty = df_prods.loc[df_prods['department_id'].isin([5, 20, 7, 12])]
df_dinnerparty.head()

08. It's important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?

# dinnerparty dimensions
df_dinnerparty.shape

# dinnerparty has 7650 rows

09. Someone from the data_engineers team in Instacart thinks they've spotted something strange about the customer with a "user_id" of 1. Extract all the information you can about this user.

df_user1 = df_ords[df_ords['user_id']==1]
df_user1

# df_user1 = df_ords.loc[df_ords['user_id'].isin([1])]

10. You also need to provide some detials about this user's behavior. What basic stats can you provide based on the information you have?

df_user1.describe()

# User 1 has made 11 purchases with an average of 6 items per purchase. 
# They purchase at an average of around every 2.5 - 3 weeks and generally shop at hours 7 - 16. 

12. Export your df_ords dataframe as "orders_wrangled.csv" in your "Prepared Data" folder.

# export orders_wrangled to prepared data folder
df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'))

13. Export the df_dep_t_new dataframe as "department_wrangled.csv" in your "Prepared Data" folder.

# export department_wrangled to prepared data folder
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'department_wrangled.csv'))
