==> 02 Scripts/04. Data Consistency Checks.ipynb <==
# Libraries

import pandas as pd
import numpy as np
import os

# Import Data

01. Import data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# products csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# wrangled order csv
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)


02. Create test data to practice on

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

# Data Consistency Prep

## df_prods

df_prods.describe()

# all id columns can be updated to string data type

price_high_outliers = df_prods[df_prods['prices'] >= 25]

price_high_outliers

# Row 21554 & 33664 look like they have errors for their prices

df_prods.shape

# Delete 2 outliers
df_prods.drop([21554, 33666], inplace = True)

df_prods.shape

01. Mixed-Type Data

# check if df_test has any mixed-type columns
for col in df_test.columns.tolist():
    weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_test[weird]) > 0:
        print(col)

# fix mixed data type found by settng the 'mix' column to string
df_test['mix'] = df_test['mix'].astype('str')

02. Missing Values

# returns number of missing data by column
df_prods.isnull().sum()

# creates new df and stores the missing values in it
df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

# run the shape function to compare the number of missing rows to overall rows
df_prods.shape

# create a new df with only rows that have no missing data i.e. "remove" missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]
# run shape to check that number of rows have decreased, thus successfully removing values
df_prods_clean.shape

03. Duplicates

# creates a subset containing only the full duplicates from df_prods_clean
df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

# create df without the duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups.shape

## df_ords

# (QUESTION 2) 
df_ords.head()
# df_ords.tail()

# (QUESTION 2)
df_ords.describe()

# The Unnamed column should be deleted - it looks like the index column was duplicated
# order_id, user_id, and order_number should be changed to data type string

01. Mixed-Type Data

# (QUESTION 3 & 4)
# check if df_ords has any mixed-type columns
for col in df_ords.columns.tolist():
    weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_ords[weird]) > 0:
        print(col)
    else:
        print(col, 'not a mixed-type')

# There are no mixed-type data columns in df_ords

02. Missing Values

# (QUESTION 5 & 6)
# returns number of missing data by column in df_ords
df_ords.isnull().sum()

# There are 206209 missing values in the days_since_prior order column.
# These could indicate 1 time customers who, therefore, have not had a prior order.
# I will leave these cells as is since the missing values indicates a 1 time customer.

03. Duplicates

# (QUESTION 7)
# creates a subset containing only the full duplicates from df_ords
df_ords_dups = df_ords[df_ords.duplicated()]
df_ords_dups

# (QUESTION 8)
# There are no duplicates, no action taken.

df_prods_clean_no_dups.shape

df_ords.shape

# Export Data

# export cleaned prods
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

# (QUESTION 9)
# export cleaned prods
df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'))
