==> 02 Scripts/05. Merge Orders and Orders_Prior.ipynb <==
# COMBINING & EXPORTING DATA

## Import Libraries

import pandas as pd
import numpy as np
import os

## Import Data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# ords
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = 0)

# ords_prior
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

## EDA

01. Head

# ords
df_ords.head()

# delete Unnamed:0 column
df_ords = df_ords.drop(columns = ['Unnamed: 0'])
df_ords.head()

# df_rds_prior
df_ords_prior.head()

02. Shape

# ords dim
df_ords.shape

# ords_prior dim
df_ords_prior.shape

## Merge DF

# merge data
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

# check if full match - does not truely check for a full match, must indicate how = 'outer'
# for this task, can leave as 'inner'
df_merged_large['_merge'].value_counts()

df_merged_large.shape

## Export as pkl 

df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))
