==> 02 Scripts/06. Merge Orders-Orders_Prior with Products.ipynb <==
# COMBINING & EXPORTING DATA PT 2

## Importing Libraries

import pandas as pd
import numpy as np
import os

## Importing Data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# products
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = 0)

# (QUESTION 3)
# orders_products_combined
df_ords_prior = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

## Analysis

01. EDA

# (QUESTION 4)
df_ords_prior.shape

df_ords_prior.head()

# delete the _merge column
df_ords_prior = df_ords_prior.drop(columns = ['_merge'])
df_ords_prior.head()

df_prods.head()

**Note:** The two data frames share the column: 'product_id'.

02. Merges

# (QUESTION 5)
# merge df_ords_prior with df_prods (the checked & cleaned version of products.csv)
df_merged_ords_prods = df_ords_prior.merge(df_prods, on = 'product_id', how = 'inner', indicator = True)

df_merged_ords_prods.head()

# (QUESTION 6)
# confirm results of the merge using the merge flag
df_merged_ords_prods['_merge'].value_counts()

df_merged_ords_prods[df_merged_ords_prods['_merge'] == 'left_only']

df_prods.loc[df_prods['product_id'].isin ([6799, 34, 116, 2240, 1511])]

df_prods.tail()

**Note:**  _merge value counts returned 30200 values with 'left_only'. After investigating some of those, it appears that there are some product_ids listed in the orders that are not in the products list.

df_merged_ords_prods.shape

## Exporting Data

# (QUESTION 7)
# export the newly merged ords_prior and prods
df_merged_ords_prods.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))
