==> 02 Scripts/07. Deriving New Variables.ipynb <==
# DERIVING NEW VARIABLES

## Importing Libraries

import pandas as pd
import numpy as np
import os
import seaborn as sns
import matplotlib.pyplot as plt

## Importing Data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# ords_prods_merge
df = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

## Define Variables

Style Guide

# chart style
sns.set_style('darkgrid')

# colors
green = '#3CBC31'
light_green = '#7CC575'
red = '#D7BBB4'
gray = '#C1C7C9'

## Analysis

01. Flagging products by price

- low range: < 5
- mid-range: 5 - 15
- high-range: > 15

**User-defined Function Method**

# defines the function that labels rows based on their price
def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'
    elif row ['prices'] > 15:
        return 'High-range product'
    else: return 'Not enough data'

# applies & runs the func to the df by creating a row called price_range to store the label
df['price_range'] = df.apply(price_label, axis = 1)

# see the counts for each range
df['price_range'].value_counts(dropna = False)

# check what the top price is
df['prices'].max()

**Loc() Pre-defined Function Method**

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# counts per category
df['price_range_loc'].value_counts(dropna = False)

df_pricesabove = df[df['prices']>14.8]
df_pricesabove

df['price_range_loc'].value_counts(dropna = False)

**For Loop Method**

We will create a new column that summarizes how busy each day of the week is.

# Find out on which day most orders take place
orders_per_day = df['orders_day_of_week'].value_counts(dropna = False)
orders_per_day

# 0 - Saturday
# 1 - Sunday
# 2 - Monday
# 3 - Tuesday
# 4 - Wednesday

# creates an empty list in which to place the results from the for loops
result = []

for value in df["orders_day_of_week"]:
    # using 0 bc using the frequncy function above, found that 0 or Saturday, is the busiest day
    if value == 0:
        result.append("Busiest day")
    # 4 or Wednesday was found to be the least busy day
    elif value == 4:
        result.append("Least busy")
    else:
        result.append("Regularly busy")

result[:5]

# combine result list with df by setting a new col equal to it
df['busiet_day'] = result

df['busiet_day'].value_counts(dropna = False)

# TASK

**Busiest Days Column** *stores the 2 busiest days and two slowest days*

# (QUESTION 2)
# Busiest days = 0 Saturday & 1 Sunday
# Slowest Days = 4 Wednesday & 3 Tuesday

# empty list to place results
top_least_days = []

for value in df["orders_day_of_week"]:
    if value == 0 or value == 1:
        top_least_days.append("Busiet days")
    elif value == 4 or value == 3:
        top_least_days.append("Least busy days")
    else:
        top_least_days.append("Regularly busy")

top_least_days[:5]

# combine resulting list with df
df['busiest_days'] = top_least_days

# (QUESTION 3)
# check values frequency of new col
df['busiest_days'].value_counts(dropna = False)

**Busiest Hours of Day Column** *stores busiest time periods od day*

# determine which col to use
df.head(1)

# will use order_hour_of_day

# detemine top and least hours of day
orders_per_hour = df['order_hour_of_day'].value_counts(dropna = False)
orders_per_hour

- **Most orders**: 10 - 15
- **Fewest orders**: 0 - 5

# (QUESTION 4)
# label top 5 as 'most orders', bottom 5 as 'fewest orders', and remaining 'average orders'

hour_business = []

for value in df["order_hour_of_day"]:
    if value >= 0 and value <= 5:
        hour_business.append("Fewest orders")
    elif value >= 10 and value <= 15:
        hour_business.append("Most orders")
    else:
        hour_business.append("Average orders")

# combine results list to df
df['busiest_period_of_day'] = hour_business

# (QUESTION 5)
# show frquency of hour business categories
df['busiest_period_of_day'].value_counts(dropna = False)

df.shape

## Export Data

# export new df created
df.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'labeled_ords_prods.pkl'))
