==> 02 Scripts/08. Grouping & Aggregating Data (Frequency, Spending, Loyalty).ipynb <==
# 4.8 TASK:
## GROUPING & AGGREGATING DATA

## Importing libraries

import pandas as pd
import numpy as np
import os
import seaborn as sns
import matplotlib.pyplot as plt

## Importing Data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# labeled_ords_prods
ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'labeled_ords_prods.pkl'))

## Defining Variables

Style Guide

# chart style
sns.set_style('darkgrid')

# colors
green = '#3CBC31'
light_green = '#7CC575'
red = '#D7BBB4'
gray = '#C1C7C9'

## Analysis

ords_prods.shape

ords_prods.head(3)

# (QUESTION 2)
# find the aggregated mean of the order_number grouped by dep id
ords_prods.groupby('department_id').agg({'order_number':['mean']})

# another syntax
# ords_prods.groupby('department_id')['order_number'].mean()

ords_prods.groupby('department_id').agg({'prices':['mean']})

**------  LOYALTY FLAG  ---------------------**

Loyalty categories are as follows:
- If max orders made by user > 40 : "Loyal customer"
- If max orders > 10 and <= 40: "Regular customer"
- If max orders <= 10: "New customer"

# creates a new col that shows a user id's total num of orders
ords_prods['max_order'] = ords_prods.groupby(['user_id'])['order_number'].transform(np.max)
ords_prods.head(5)

# assign loyalty flags in a new col caled loyalty-flag
# Loyal
ords_prods.loc[ords_prods['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

# Regular
ords_prods.loc[(ords_prods['max_order'] > 10 ) & (ords_prods['max_order'] <= 40), 'loyalty_flag'] = 'Regular customer'

# New
ords_prods.loc[ords_prods['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Number breakdown of each category
ords_prods['loyalty_flag'].value_counts(dropna = False)

# (QUESTION 5)
# show average price of products purchased by loyalty flag
ords_prods.groupby('loyalty_flag').agg({'prices':['mean']})

# spend about the same per product

**------   SPENDING FLAG   ---------------------**

Spending categories are as follows:
- If mean price of products purchased by user >= 10 : 'High spender'
- If mean price < 10 : 'Low spender'

# (QUESTION 6)
# create new col that shows a user's avg per product
ords_prods['avg_spent'] = ords_prods.groupby(['user_id'])['prices'].transform(np.mean)
ords_prods.head(5)

# assign labels for newly created avg per product
# High spender
ords_prods.loc[ords_prods['avg_spent'] >= 10, 'spending_flag'] = 'High spender'

# Low spender
ords_prods.loc[ords_prods['avg_spent'] < 10, 'spending_flag'] = 'Low spender'

ords_prods['spending_flag'].value_counts(dropna = False)

**------   FREQUENCY FLAG   ---------------------**

Frequency categories are as follows:
- If median days_since_prior_order > 20 : 'Non-frequent customer'
- If median days > 10 & <= 20 : 'Regular customer'
- If median days <= 10 : 'Frequent customer'

# (QUESTION 7)
# create new col that shows a user's median days_since_prior_order
ords_prods['med_since_prior'] = ords_prods.groupby(['user_id'])['days_since_prior_order'].transform(np.median)
ords_prods.head(5)

# assign labels for newly created med days since
# Non-frequent
ords_prods.loc[ords_prods['med_since_prior'] > 20, 'frequency_label'] = 'Non-frequent customer'

# Regular customer
ords_prods.loc[(ords_prods['med_since_prior'] > 10) & (ords_prods['med_sinceprior'] <= 20), 'frequency_label'] = 'Regular customer'

# Frequnt customer
ords_prods.loc[ords_prods['med_since_prior'] <= 10, 'frequency_label'] = 'Frequent customer'

ords_prods['frequency_label'].value_counts(dropna = False)

## Export Data

# export df
ords_prods.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_wflags.pkl'))
