==> 02 Scripts/09. Clean Customers & Merge to DF.ipynb <==
# 4.9 Task

## Importing Libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

## Importing Data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# customer
cust_df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

# ords_prods
ords_prods = pd.read_pickle(os.path.join(path, '02 data', 'Prepared Data', 'ords_prods_wflags.pkl'))

## Analysis

cust_df.head()

cust_df.info()

**Cleaning Data**

01. Rename Columns

# First Name -> first name
cust_df.rename(columns = {'First Name': 'first name'}, inplace = True)
# Surname -> Last Name
cust_df.rename(columns = {'Surnam': 'last name'}, inplace = True)
# Gender -> gender
cust_df.rename(columns = {'Gender': 'gender'}, inplace = True)
# STATE -> State
cust_df.rename(columns = {'STATE': 'state'}, inplace = True)
# Age -> age
cust_df.rename(columns = {'Age': 'age'}, inplace = True)
# n_dependants -> dependants
cust_df.rename(columns = {'n_dependants': 'dependents'}, inplace = True)

cust_df.head(3)

cust_df.shape

**Quality & Consistency Checks**

01. Mixed-Type Data

for col in cust_df.columns.tolist():
    weird = (cust_df[[col]].map(type) != cust_df[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (cust_df[weird]) > 0:
        print(col)

# set first name to type string
cust_df['first name'] = cust_df['first name'].astype('str')

02. Missing Values

cust_df.isnull().sum()

# no missing values

03. Duplicates

df_dups = cust_df[cust_df.duplicated()]
df_dups.head()

# no full duplicates

## Merge Data

01. Identify shared column

cust_df.head(1)

ords_prods.head(1)

02. Execute merge

ords_prods_cust = ords_prods.merge(cust_df, on = 'user_id')

ords_prods_cust.head(3)

ords_prods_cust.shape

## Export Data

ords_prods_cust.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_cust.pkl'))
