==> 02 Scripts/11. Excel Crosstab.ipynb <==
# Excel Reporting

## Importing Libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

## Importing Data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# ords_prods_wflags
ords_prods_cust = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_cust.pkl'))

## Analysis

**Crosstabs** : *Python's version of Excel's pivot tables*

# create pivot table of days_since_prior_order and order number
# to check of all null values in prior_order are only null for first time customers
# who would have an order number of 1
# If there are null values for other order #s, this would indicate a true missing value.

crosstab = pd.crosstab(ords_prods_cust['days_since_prior_order'], ords_prods_cust['order_number'], dropna = False)

# will be cumbersome to see full pivot table in Pyhon, so best to view in Excel
# the following copies the able to your clipboard

crosstab.to_clipboard()
