==> 02 Scripts/12. Cleaning & Reducing Memory.ipynb <==
# Clean & Reduce Memory Usage

## Importing Libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

## Importing Data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# ords_prods_cust
ords_prods_cust = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_cust.pkl'))

## Analysis

#### Cleaning

01. Clean Dataset

# set limit of column display to 40, allowing all columns in df to be seen
pd.set_option('display.max_column', 40)

ords_prods_cust.head()

Renaming Columns

# fix typo - changing orders to order in order_day_of_week
ords_prods_cust.rename(columns = {'orders_day_of_week': 'order_day_of_week'}, inplace = True)
# rename for clarity
ords_prods_cust.rename(columns = {'busiet_day': 'day_busyness'}, inplace = True)
ords_prods_cust.rename(columns = {'busiest_period_of_day': 'hour_of_day_busyness'}, inplace = True)
# rename for clarity
ords_prods_cust.rename(columns = {'med_since_prior': 'med_days_since_prior_order'}, inplace = True)

Deleting Columns

# delete unnecessary / duplicate columns
opc_clean = ords_prods_cust.drop(columns = ['_merge', 'price_range', 'med_sinceprior'])

opc_clean.head(1)

02. Remove PII

# [first name] and [last name]  are PII - these columns will be deleted
opc_nopii = opc_clean.drop(columns = ['first name', 'last name'])

opc_nopii.head(1)

03. Reduce Memory Usage

opc_nopii['order_id'] = opc_nopii['order_id'].astype('int32')
opc_nopii['user_id'] = opc_nopii['user_id'].astype('int32')
opc_nopii['order_number'] = opc_nopii['order_number'].astype('int8')
opc_nopii['order_day_of_week'] = opc_nopii['order_day_of_week'].astype('int8')
opc_nopii['order_hour_of_day'] = opc_nopii['order_hour_of_day'].astype('int8')
opc_nopii['days_since_prior_order'] = opc_nopii['days_since_prior_order'].astype('float16')

# Export Data

opc_nopii.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'opc_checked.pkl'))
