==> 02 Scripts/13. Customer Profile Flags.ipynb <==
## Importing Libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

## Importing Data

# main folder
path = r'C:\Users\steve\Documents\11.24 Instacart Basket Analysis'

# ords_checked
opc = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'opc_checked.pkl'))

## Analysis

#### 01. Regions

# create lists for each region
northeast = ['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey']
midwest = ['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri']
south = ['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana']
west = ['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 'California', 'Hawaii']

# define function that will assign region based on [state] column
result = []

for value in opc['state']:
    if value in northeast:
        result.append('northeast')
    elif value in midwest:
        result.append('midwest')
    elif value in south:
        result.append('south')
    elif value in west:
        result.append('west')
    else: 
        result.append('error')

# set a new column in df equal to result
opc['region'] = result

opc['region'].value_counts(dropna = False)

# the following were used to troubleshoot when had values appearing with the 'error' tag
# ----------
#region_error = opc[opc['region'] == 'error']
#region_error['state'].value_counts(dropna = False)

#### 02. Spending Habits by Region

# show the counts of spending_flag for each region
spender_count_by_region = opc.groupby(['region'])['spending_flag'].agg(pd.Series.count)
spender_count_by_region

# show the counts of each value with spending_flag separated by region
spender_flag_count = opc.groupby(['region'])['spending_flag'].value_counts(dropna = False)

spender_flag_count

--------------------------------------------------------------
##### **Insight:** *All regions top spender type is 'low spender'*

--------------------------------------------------------------

# convert spender_flag counts into percentages
spend_percent = spender_flag_count.groupby(level=0).apply(lambda x: 100 * x / float(x.sum())).round(2).reset_index(level=0, drop=True)
# spend_percent

# convert series to a dataframe
spend_percent_df = spend_percent.to_frame()

# assign column name 'percent'
spend_percent_df.rename(columns = {'count': 'percent'}, inplace = True)
spend_percent_df

--------------------------------------------------------------
##### **Insight:** *All regions have about 99.6 % low spenders and about .4% high spenders.*

--------------------------------------------------------------

#### 03. Exlusion flag for low activity customers

- **Low activity customers** = customers with < 5 orders

pd.options.display.max_columns = 35
opc.head(1)

# create a new df that stores all customers with a max order number less than 5
low_activity_cust = opc[opc['max_order'] < 5]
# show the count of customers per order number less than 5
low_activity_cust['max_order'].value_counts(dropna = False)

opc.shape

low_activity_cust.shape

# df excluding low activity should be 30959687 rows

# create a df that excludes these low activity customers
opc_minus_low_activity_cust = opc[opc['max_order'] >= 5]
opc_minus_low_activity_cust.shape

#### 04. Customer Profiles

**Age**

# for simplicity, rename df
df = opc_minus_low_activity_cust

# find the age range of df using min and max
df['age'].min()

df['age'].max()

Age categories
- **young adult**: age < 25
- **adult**: age >= 25 and < 40
- **middle-aged**: age >= 40 and < 60
- **senior**: age >= 60

# define the function that labels ros based on the age group conditions
def age_label(row):
    if row['age'] < 25:
        return 'young adult'
    elif (row['age'] >= 25) and (row['age'] < 40):
        return 'adult'
    elif (row['age'] >= 40) and (row['age'] < 60):
        return 'middle-aged'
    elif row['age'] >= 60:
        return 'senior'
    else: return 'error'

# create new row for age group using the age_label function
df['age_group'] = df.apply(age_label, axis = 1)

# see the counts for each age range
df['age_group'].value_counts(dropna = False)

Income categories

df['income'].min()

df['income'].max()

Income categories
- **lower class**: income < 53,740
- **middle class**: income >= 53, 740 and <= 161,220 (cnbc definition of middle class in 2023)
- **upper class**: income > 161,220

# define the function that labels rows based on the income group conditions
def income_label(row):
    if row['income'] < 53740:
        return 'lower class'
    elif (row['income'] >= 53740) and (row['income'] <= 161220):
        return 'middle class'
    elif row['income'] > 161220:
        return 'upper class'
    else: return 'error'

# create new row for income group using the income_label function
df['income_group'] = df.apply(income_label, axis = 1)

# see the counts for each age range
df['income_group'].value_counts(dropna = False)

Number of Dependents

Dependents flag:
- **No dependents**: dependents = 0 
- **Has dependents**: dependents > 0

# define the function that labels rows based on the income group conditions
def dependents_label(row):
    if row['dependents'] == 0:
        return 'No dependents'
    elif row['dependents'] > 0:
        return 'Has dependents'
    else: return 'error'

# create new row for age group using the age_label function
df['dependents_flag'] = df.apply(dependents_label, axis = 1)

# see the counts for each age range
df['dependents_flag'].value_counts(dropna = False)

## Export Data

df.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_all.pkl'))
